move genClient -> adapter-next
